Create eula.go
